// src/data/store/reducer/types.ts
// eslint-disable-next-line no-shadow
export enum ReducerTypes {
  changeState = '@app/changeState',
  awaitAnswer = '@app/awaitAnswer',
  startGame = '@app/startGame',
  startVideo = '@app/startVideo',
  answer = '@app/answer',
  setPlaylist = '@app/setPlaylist',
  clearAnswer = '@app/clearAnswer',
  reset = '@app/reset',
}